backend/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nba-api==1.4.1
numpy==1.24.4
pandas==2.0.3
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.30
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.18.1
